+++ src/lib.rs
@@ -1,4 +1,5 @@
 mod archive;
+mod async_server;
 mod broker;
 mod client;
 mod complete;
@@ -19,6 +20,9 @@ mod state;
 mod task;
 
 pub use archive::{ArchiveBroker, ArchiveError};
+pub use async_server::{
+    AsyncServer, AsyncSleeper, AsyncWorkerProcessor, DEFAULT_ASYNC_SERVER_IDLE_SLEEP, TokioSleeper,
+};
 pub use broker::redis::{
     DEFAULT_LEASE_DURATION, RedisArchivePlan, RedisArchivePlanError, RedisArg, RedisBroker,
     RedisBrokerError, RedisClientExecutor, RedisCommandExecutor, RedisCompletePlan,